requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
pdf2image==1.16.3
Pillow==10.1.0
pandas==2.1.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
aiofiles==23.2.1
python-dotenv==1.0.0

